+++ Core/Util/RandomGenerators/NoRandom.cs
@@ -2,7 +2,7 @@
 {
     public class NoRandom : IRandom
     {
-        public byte NextByte() => 0;
+        public int NextByte() => 0;
         public int NextDiff() => 0;
         public int RandomIndex => 0;
